requirements.txt
# ============================================================================

# CORE UI & DATA

# ============================================================================

streamlit==1.29.0

pandas==2.1.4

numpy==1.26.3

requests==2.31.0

python-dotenv==1.0.0

plotly==5.18.0
 
# ============================================================================

# SNOWFLAKE

# ============================================================================

snowflake-connector-python[pandas]==3.7.0

sqlparse==0.5.0
 
# ============================================================================

# AI & AGENTS

# ============================================================================

anthropic>=0.75.0

langchain==0.3.0

langchain-anthropic==0.3.0

langchain-community==0.3.0

langgraph==0.2.45
 
# ============================================================================

# RAG (Vector Store)

# ============================================================================

pinecone-client==3.0.0

# Only install the CPU version of torch to save massive space

--extra-index-url https://download.pytorch.org/whl/cpu

sentence-transformers==3.0.0